Converted Gonk into a static library. Cleaned up warnings.

+++ Gonk/src/Platform/Windows/WindowsInput.cpp
@@ -42,7 +42,7 @@ namespace Gonk {
             GLFWwindow* window = (GLFWwindow*)Application::Get().GetWindow().GetNativeWindow();
             double xpos, ypos;
             glfwGetCursorPos(window, &xpos, &ypos);
-            return std::pair<float, float>(xpos, ypos);
+            return std::pair<float, float>((float)xpos, (float)ypos);
         }
     };
